requirements.txt
openai
pydub
python-dotenv
google-cloud-texttospeech
flask 
flask_restful